# docs/openapi.yaml
openapi: 3.0.3
info:
  title: HalalPilot Orchestration API
  version: 0.1.0
  description: |
    Kontrak antara skill OpenClaw `halalpilot` (pemanggil) dan layanan Express+SQLite (pelaksana).
    Prinsip: agent memutuskan bahasa & alur percakapan; API memutuskan hukum (rules/*.yaml), status, dan jadwal.
    Semua endpoint di 127.0.0.1:3000, header `Authorization: Bearer ${HALALPILOT_API_TOKEN}`.
    Endpoint yang mengubah state menerima `Idempotency-Key` (disarankan skill mengisi `<aksi>:<umk>:<hash payload>`).
servers:
  - url: http://127.0.0.1:3000/api/v1
security:
  - bearer: []
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
  schemas:
...
    Ingredient:
      type: object
      properties:
        nama_asli: { type: string }
        nama_normal: { type: string, nullable: true }
        kelas: { type: string, enum: [dikecualikan_A, dikecualikan_B, dikecualikan_C, positif, kritis, tidak_dikenal] }
        rule_id: { type: string }
        butuh_sertifikat_pemasok: { type: boolean }
        dikonfirmasi_umk: { type: boolean }
...
paths:
...
  /products/{pid}:
    patch:
      summary: Ubah atribut produk dari keterangan UMK (cerita proses produksi, jumlah teknik pengawetan, giling sendiri, jenis). Menyimpan lalu mengevaluasi ulang otomatis.
      parameters: [{ name: pid, in: path, required: true, schema: { type: integer } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                proses_ringkas: { type: string, minLength: 20, maxLength: 1000 }
                teknik_pengawetan_count: { type: integer, minimum: 0, maximum: 5 }
                giling_sendiri: { type: integer, enum: [0, 1] }
                jenis: { type: string }
      responses: { '200': { description: "{product_id, ...field, decision}" } }
  /products/{pid}/media:
    post:
      summary: Simpan foto (label/proses) dari inbound OpenClaw; mengembalikan media_id
      parameters: [{ name: pid, in: path, required: true, schema: { type: integer } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [kind, source_path]
              properties:
                kind: { type: string, enum: [label, proses, sertifikat_pemasok] }
                source_path: { type: string, description: "path file media inbound di ~/.openclaw (disalin, di-hash)" }
      responses: { '201': { description: "{media_id, sha256, dokumen_diterima, chase_dibatalkan}. kind=label → FOTO_PRODUK diterima; kind=proses → PROSES diterima (otomatis, sisa pengejaran dokumen itu dibatalkan)" } }
  /products/{pid}/ingredients:
    put:
      summary: |
        Simpan daftar bahan hasil ekstraksi vision (oleh agent) ATAU hasil koreksi UMK.
        API menormalisasi & mengklasifikasi tiap bahan (rules/ingredients.yaml). Belum memutuskan jalur.
      parameters: [{ name: pid, in: path, required: true, schema: { type: integer } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [bahan, sumber]
              properties:
                bahan: { type: array, items: { type: string }, example: ["tepung terigu", "margarin", "perisa vanila", "telur"] }
                sumber: { type: string, enum: [vision, umk_koreksi] }
                dikonfirmasi_umk: { type: boolean, default: false }
      responses:
        '200':
          description: Klasifikasi per bahan
          content:
            application/json:
              schema:
                type: object
                properties:
                  ingredients: { type: array, items: { $ref: '#/components/schemas/Ingredient' } }
                  ringkasan: { type: string, example: "4 bahan: 1 dikecualikan, 1 positif, 2 kritis (butuh sertifikat pemasok)" }
                  perlu_konfirmasi: { type: boolean }
  /products/{pid}/extract:
    post:
      summary: |
        FALLBACK V1: ekstraksi bahan dari foto label oleh API (memanggil OpenRouter vision langsung).
        Dipakai hanya jika imageModel OpenClaw tidak menerima foto Telegram.
      parameters: [{ name: pid, in: path, required: true, schema: { type: integer } }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { media_id: { type: integer } } } } } }
      responses: { '200': { description: "{bahan[], proses_ringkas, confidence}" } }
...